tests: combine send message test files
